+++ api/index.cjs
@@ -4,25 +4,25 @@ const cors = require("cors");
 const { serverPort } = require("./constatns.cjs");
 const app = express();
 const http = require("http").Server(app);
-const io = require("socket.io")(http);
+// const io = require("socket.io")(http);
 
 app.use(cors());
 app.use(express.json());
 
 //  -----------------------------------
 // TODO SOCKETS
-io.on("connection", (socket) => {
-  console.log("User Connected");
-  socket.join("folders");
-  console.log(socket.folders);
+// io.on("connection", (socket) => {
+//   console.log("User Connected");
+//   socket.join("folders");
+//   console.log(socket.folders);
 
-  socket.on("folders", (message) => {
-    console.log(message);
-  });
-  socket.on("disconnect", () => {
-    console.log("User Disconnected");
-  });
-});
+//   socket.on("folders", (message) => {
+//     console.log(message);
+//   });
+//   socket.on("disconnect", () => {
+//     console.log("User Disconnected");
+//   });
+// });
 //  -----------------------------------
 
 app.get("/addFolder", (req, res) => {
@@ -32,6 +32,7 @@ app.get("/addFolder", (req, res) => {
       props: {
         name: name,
         id: id,
+        type: "folder",
       },
       table: "folders",
     });
@@ -51,13 +52,18 @@ app.get("/getFolders", async (req, res) => {
     res.status(400).send("error");
   }
 });
-app.get("/deleteFolder", async (req, res) => {
+app.delete("/folder", async (req, res) => {
   try {
-    const select = await surrealDB.delete({
-      id: `folders:${req.query.id}`,
-    });
-    res.status(200).send(await select[0].result);
+    const id = String(req.query.id);
+    const select = await surrealDB
+      .delete({
+        id: id.includes("folders") ? id : `folders:${id}`,
+      })
+      .then(() => res.status(200).send(true))
+      .catch(() => res.status(400).send("error"));
+    console.log(select);
   } catch (error) {
+    console.log(error);
     res.status(400).send("error");
   }
 });