+++ src/CenterPanel/CenterPanel.tsx
@@ -4,12 +4,11 @@ import { useSelector } from "react-redux";
 import { getCurrentFolder, getFolderItems } from "../redux/selectors/folders";
 import EmptyState from "./components/EmptyState/EmptyState";
 import Header from "./components/Header/Header";
-import './ConterPanel.scss'
+import './CenterPanel.scss'
 const CenterPanel = () => {
   const currentFolder = useSelector(getCurrentFolder)
   const folderItems = useSelector(getFolderItems)
   
-  console.log(folderItems)
   return (
     <div className="center-panel">
       <Header title={currentFolder?.name} />