+++ src/icons/RadioButton.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+
+const RadioButton = () => {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M12 24C18.6274 24 24 18.6274 24 12C24 5.37258 18.6274 0 12 0C5.37258 0 0 5.37258 0 12C0 18.6274 5.37258 24 12 24ZM12 22C6.47716 22 2 17.5228 2 12C2 6.47716 6.47716 2 12 2C17.5228 2 22 6.47716 22 12C22 17.5228 17.5228 22 12 22Z"
+        fill="white"
+        fillOpacity="0.85"
+      />
+    </svg>
+  );
+};
+export default RadioButton;