+++ changelog.md
@@ -1,5 +1,52 @@
 ## Changelogs List
 
+## 0.0.4 - Altaev Nurzhan
+
+### About
+
+- Function to detete "folders"
+- Added fonts, colors as variables
+- Added icons
+- Left Panel style changes
+- Added request to remove folder
+- Fix preventDefault in browsers
+### File Logs
+
+#### Added: 
+- src/icons/Archive.tsx
+- src/icons/Calendar.tsx
+- src/icons/CalendarToday.tsx
+- src/icons/More.tsx
+- src/icons/Plus.tsx
+- src/icons/RadioButton.tsx
+- src/icons/RadioButtonSelected.tsx
+- src/icons/Settings.tsx
+- src/icons/Storage.tsx
+- src/icons/index.ts
+- src/components/Checkbox/Checkbox.scss
+- src/components/Checkbox/Checkbox.tsx
+- src/fonts.scss
+#### Added: 
+- api/index.cjs
+- api/surreal.cjs
+- App.tsx
+- LeftPanel/LeftPanel.module.scss
+- LeftPanel/LeftPanel.tsx
+- LeftPanelInput.module.scss
+- LeftPanelInput.tsx
+- LeftPanelItems.module.scss
+- LeftPanelItems.tsx
+- ContextItems.tsx
+- ContextMenu.tsx
+- fetch.ts
+- hooks/ContextMenuHandler.tsx
+- hooks/useLocale.tsx
+- index.scss
+- index.tsx
+- redux/index.tsx
+- redux/selectors/folders.tsx
+- redux/store/FoldersStore.tsx
+---
 ## 0.0.31 - Altaev Nurzhan
 
 ### About
@@ -7,7 +54,7 @@
 - Fix ContextMenu positions
 - Added custom Context Menu for page
 - Small Left Panel style changes
-
+---
 ### File Logs
 
 #### Added: 
@@ -31,7 +78,7 @@
 - src/index.tsx
 - tsconfig.json
 - webpack.config.babel.ts
-
+---
 ## 0.0.3 - Altaev Nurzhan
 
 ### About
@@ -57,6 +104,7 @@
 - src/components/ContextMenu/ContextMenu.scss
 - src/components/ContextMenu/ContextMenu.tsx
 - src/redux/store/FoldersStore.tsx
+---
 ## 0.0.2 - Altaev Nurzhan
 
 ### About
@@ -91,3 +139,4 @@
 - src/LeftPanel/components/LeftPanelItems/LeftPanelItems.module.scss
 - src/LeftPanel/components/LeftPanelItems/LeftPanelItems.tsx
 - tsconfig.json
+---